Refactored some names + added setup function to ProvingSystemUtil trait
