+++ cctp_primitives/src/utils/serialization.rs
